gym_requirements.txt
cogment[generate]==2.7.1
grpcio==1.43.0
hydra-core==1.3.2
mlflow==2.2.2
names_generator~=0.1.0
opencv-python~=4.5.5
PyYAML~=5.3.1
starlette==0.20.0
uvicorn==0.17.6
boto3==1.26.123

# environments
gym[box2d,classic_control, mujoco]>=0.25,<0.26
PettingZoo[atari, classic]~=1.22.2,<1.23
tinyscaler~=1.2.4
Gymnasium~=0.26.3
AutoROM~=0.4.2


# actors
torch==1.11.0
numpy>=1.21.5,<1.22

# For testing
black~=22.3.0
pylint~=2.14
pytest~=6.2.5
pytest-benchmark~=4.0.0
pytest-timeout~=2.1.0

# For Mlflow
alembic==1.9.3
SQLAlchemy<2
